+++ ImageFlow/IFBorderlessWindow.m
@@ -12,7 +12,7 @@
 @implementation IFBorderlessWindow
 
 - (id)initWithContentRect:(NSRect)contentRect
-                styleMask:(unsigned int)styleMask
+                styleMask:(NSUInteger)styleMask
                   backing:(NSBackingStoreType)bufferingType
                     defer:(BOOL)deferCreation;
 {
